chore: extract handleCommonErrors to a separate module
clean up index.ts a bit
